+++ example/lib/main.dart
@@ -24,13 +24,28 @@ class KakaoMapTest extends StatelessWidget {
               lng: 126.570667,
               showMapTypeControl: true,
               showZoomControl: true,
+              draggableMarker: true,
+              mapType: MapType.BICYCLE,
+              polygon: KakaoPolygon(
+                polygon: [
+                  KakaoLatLng(33.45086654081833, 126.56906858718982),
+                  KakaoLatLng(33.45010890948828, 126.56898629127468),
+                  KakaoLatLng(33.44979857909499, 126.57049357211622),
+                  KakaoLatLng(33.450137483918496, 126.57202991943016),
+                  KakaoLatLng(33.450706188506054, 126.57223147947938),
+                  KakaoLatLng(33.45164068091554, 126.5713126693152)
+                ],
+                polygonColor: Colors.red,
+                polygonColorOpacity: 0.3,
+                strokeColor: Colors.deepOrange,
+                strokeWeight: 2.5,
+                strokeColorOpacity: 0.9
+              ),
               markerImageURL:
                   'https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/marker_red.png',
-              onTapMarker: (message) async {
+              onTapMarker: (message) {
                 ScaffoldMessenger.of(context)
-                    .showSnackBar(SnackBar(content: Text('Marker is clicked')));
-
-                //await _openKakaoMapScreen(context);
+                    .showSnackBar(SnackBar(content: Text(message.message)));
               }),
           ElevatedButton(
               child: Text('Kakao map screen'),
@@ -54,5 +69,46 @@ class KakaoMapTest extends StatelessWidget {
 
     Navigator.push(
         context, MaterialPageRoute(builder: (_) => KakaoMapScreen(url: url)));
+  }
+
+  Widget _testingCustomScript(
+      {required Size size, required BuildContext context}) {
+    return KakaoMapView(
+        width: size.width,
+        height: 400,
+        kakaoMapKey: kakaoMapKey,
+        lat: 33.450701,
+        lng: 126.570667,
+        customScript: '''
+    var markers = [];
+    
+    function addMarker(position) {
+    
+      var marker = new kakao.maps.Marker({position: position});
+
+      marker.setMap(map);
+    
+      markers.push(marker);
+    }
+    
+    for(var i = 0 ; i < 3 ; i++){
+      addMarker(new kakao.maps.LatLng(33.450701 + 0.0003 * i, 126.570667 + 0.0003 * i));
+      
+      kakao.maps.event.addListener(markers[i], 'click', function(){
+        onTapMarker.postMessage('marker ' + i + ' is tapped');
+     });
+    }
+    
+		  var zoomControl = new kakao.maps.ZoomControl();
+      map.addControl(zoomControl, kakao.maps.ControlPosition.RIGHT);
+   
+      var mapTypeControl = new kakao.maps.MapTypeControl();
+      map.addControl(mapTypeControl, kakao.maps.ControlPosition.TOPRIGHT);
+    
+              ''',
+        onTapMarker: (message) {
+          ScaffoldMessenger.of(context)
+              .showSnackBar(SnackBar(content: Text(message.message)));
+        });
   }
 }