+++ lib/kakaomap_webview.dart
@@ -3,3 +3,6 @@ library kakaomap_webview;
 export 'src/kakaomap_view.dart';
 export 'src/kakaomap_util.dart';
 export 'src/kakaomap_screen.dart';
+export 'src/kakaomap_type.dart';
+export 'src/kakao_latlng.dart';
+export 'src/kakao_polygon.dart';