+++ lib/src/kakao_latlng.dart
@@ -1,0 +1,10 @@
+/// This class is used only for polygon now
+class KakaoLatLng {
+  /// Latitude
+  double lat;
+
+  /// Longitude
+  double lng;
+
+  KakaoLatLng(this.lat, this.lng);
+}